meet the requirement for 50f distance destination

+++ core/src/com/unimelb/swen30006/partc/core/World.java
@@ -92,7 +92,7 @@ public class World implements ISteppable {
 		//add a new parking car as an obstacle
 		this.cars[1] = new Car(new Point2D.Double(30,145), 6, 10, Color.BLACK, 25f, 50f, 6f );
 
-		Planner planner = new Planner(cars[0], new Point2D.Double(350, 270), new Map(roads, intersections));
+		Planner planner = new Planner(cars[0], new Point2D.Double(350, 240), new Map(roads, intersections));
 		Planner planner2 = new Planner(cars[1], new Point2D.Double(380, 500), new Map(roads, intersections));
